requirements/vk_api-requirements.txt
requests==2.28.2
aiohttp==3.9.1
requests-oauthlib==1.3.1
Pillow==9.4.0
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==1.10.7
vk-api==11.9.9
async-timeout==4.0.2
nest-asyncio==1.5.6